Extract common components of battery temps and voltage
