+++ src/ui/battery/mod.rs
@@ -1,0 +1,3 @@
+pub mod battery_temps;
+pub mod battery_voltage;
+pub mod common;
