+++ src/ui/mod.rs
@@ -1,5 +1,4 @@
-pub mod battery_temps;
-pub mod battery_voltage;
+pub mod battery;
 pub mod bootloader;
 pub mod bus_load;
 pub mod command_palette;
